get_mark_from_yaml_data: move to spack.util.spack_yaml
